UPDATE: Now the pagination is showing up properly. To "fix" the issue temporarily I did the following, BUT removing "hidden" seemed to be for handling responsive HTML output for "sm" (small) browser windows. Maybe someone working on the pagination at Laravel(?) or Tailwind(?) broke something if 'npm run build' happens to pull some online file(s)???
ORIGINALLY:
For some reason the links() pagination method had a "hidden" Tailwind CSS class in the "showing X to Y of Z results" causing it to become hidden in the browser. Not sure if something changed using some online file(s) to generate the pagination HTML or what.

To "fix" the issue, I had to:

1) publish the pagination views by running 'php artisan vendor:publish --tag=laravel-pagination'
2) open the pagination Blade file (resources/views/vendor/pagination/tailwind.blade.php)
3) remove the "hidden" class from '<div class="hidden sm:flex-1 sm:flex sm:items-center sm:justify-between">'
4) save changes
5) test

All of a sudden I noticed the "showing" text and nav were missing.